spec/inventoryUpdate.spec.ts
import {InventoryUpdate} from "../src/inventoryUpdate";

describe('InventoryUpdate', () => {
    it('should take in two arrays with first one empty and return one array in order', () => {
        expect(InventoryUpdate.update([], [[2, "Hair Pin"], [3, "Half-Eaten Apple"], [67, "Bowling Ball"], [7, "Toothpaste"]]))
            .toEqual([[67, "Bowling Ball"], [2, "Hair Pin"], [3, "Half-Eaten Apple"], [7, "Toothpaste"]]);
    });
    it('should take in two arrays with second one empty and return one array in order', () => {
        expect(InventoryUpdate.update([[2, "Hair Pin"], [3, "Half-Eaten Apple"], [67, "Bowling Ball"], [7, "Toothpaste"]], []))
            .toEqual([[67, "Bowling Ball"], [2, "Hair Pin"], [3, "Half-Eaten Apple"], [7, "Toothpaste"]]);
    });
    it('should take in two arrays and return one array combined and in order', () => {
        expect(InventoryUpdate.update([[21, "Bowling Ball"], [2, "Dirty Sock"], [1, "Hair Pin"], [5, "Microphone"]], [[2, "Hair Pin"], [3, "Half-Eaten Apple"], [67, "Bowling Ball"], [7, "Toothpaste"]]))
            .toEqual([[88, "Bowling Ball"], [2, "Dirty Sock"], [3, "Hair Pin"], [3, "Half-Eaten Apple"], [5, "Microphone"], [7, "Toothpaste"]]);
    });
    it('should take in two arrays and return one array combined and in order', () => {
        expect(InventoryUpdate.update([[21, "Bowling Ball"], [2, "Dirty Sock"], [1, "Hair Pin"], [5, "Microphone"]], []))
            .toEqual([[21, "Bowling Ball"], [2, "Dirty Sock"], [1, "Hair Pin"], [5, "Microphone"]]);
    });
});
